Create Esri FeatureSet JSON

Represent simple features as Esri JSON.

These functions convert R objects to Esri json representations. There are three types of representations. These are, from smallest to largest, a geometry object, a feature and a feature set.

Usage

st_as_features(x, ...)

st_as_featureset(x, ...)

st_as_json(x, crs, ...)

Arguments

x

an object of class sfg

...

unused

crs

a CRS ID, crs object, or a well-known text representation of CRS

Details

st_as_json() converts an sfg object to a geometry object

st_as_features() converts an sfc or sf object to a list of features

st_as_featureset() converts an sf, sfc, or data.frame to a feature set object

Geometry object contain the coordinates for a geometry. Features are geometries that have associated attributes with them. This would be similar to a row in an sf object. FeatureSets are a list of features that have additional metadata fields such as spatialReference, geomtryType, and fields. FeatureSets correspond to an sf object.

Geometry objects are defined for 5 different types. These are:

Point: esriGeometryPoint

Multipoint: esriGeometryMultipoint

Polyline: esriGeometryPolyline

note that polyline encompasses both LINESTRING and MULTILINESTRING

Polygon: esriGeometryPolygon

note that polygon encompasses both POLYGON and MULTIPOLYGON

Envelope: esriGeometryEnvelope

envelopes correspond with bounding boxes but can have a Z dimension

Examples

library(sf)
#> Linking to GEOS 3.11.0, GDAL 3.5.3, PROJ 9.1.1; sf_use_s2() is TRUE
st_as_json(st_point(c(0, 1, 3, 4)))
#> {"hasZ":true,"hasM":true,"x":0.0,"y":1.0,"z":3.0,"m":4.0,"spatialReference":{"wkid":4326}}
st_as_json(st_multipoint(x = matrix(runif(4), ncol = 2)))
#> {"hasZ":false,"hasM":false,"points":[[0.8853790711145848,0.20541643188335002],[0.19411082635633648,0.6049823090434074]],"spatialReference":{"wkid":4326}}
st_as_json(st_linestring(x = matrix(runif(2), ncol = 2)))
#> {"hasZ":false,"hasM":false,"paths":[[0.6975813247263432,0.9091675197705627]],"spatialReference":{"wkid":4326}}
st_as_json(st_linestring(x = matrix(runif(2), ncol = 2)))
#> {"hasZ":false,"hasM":false,"paths":[[0.06757986918091774,0.10631556552834809]],"spatialReference":{"wkid":4326}}

# polygon
m <- matrix(
  c(0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 2, 2, 1, 2, 3, 1, 3, 2, 0, 0, 0),
  ncol = 3,
  byrow = TRUE
)
st_as_json(st_polygon(list(m)))
#> {"hasZ":true,"hasM":false,"rings":[[[0.0,0.0,0.0],[0.0,0.0,1.0],[0.0,1.0,1.0],[1.0,2.0,2.0],[1.0,2.0,3.0],[1.0,3.0,2.0],[0.0,0.0,0.0]]],"spatialReference":{"wkid":4326}}